refactor: move prompt ui logic into printers.go
