+++ cmd/find.go
@@ -2,16 +2,12 @@ package cmd
 
 import (
 	"fmt"
-	"regexp"
-	"strings"
 
 	"github.com/elhmn/ckp/internal/config"
+	"github.com/elhmn/ckp/internal/printers"
 	"github.com/elhmn/ckp/internal/store"
-	"github.com/manifoldco/promptui"
 	"github.com/spf13/cobra"
 )
-
-const selectItemsSize = 10
 
 //NewFindCommand display a prompt for you to enter the code or solution you are looking for
 func NewFindCommand(conf config.Config) *cobra.Command {
@@ -35,59 +31,6 @@ func NewFindCommand(conf config.Config) *cobra.Command {
 	command.PersistentFlags().Bool("from-history", false, `list code and solution records from history`)
 
 	return command
-}
-
-func trimText(s string) string {
-	if len(s) > 50 {
-		return s[:50] + "..."
-	}
-	return s
-}
-
-func getTemplates() *promptui.SelectTemplates {
-	funcMap := promptui.FuncMap
-	funcMap["inline"] = func(s string) string {
-		return strings.ReplaceAll(trimText(s), "\n", " ")
-	}
-
-	//if you find a hard time understand it check out golang templating format documentation
-	//here https://golang.org/pkg/text/template
-	return &promptui.SelectTemplates{
-		Label: "{{ if .Code.Content -}} {{`code:` | bold | green}} " +
-			"{{ inline .Code.Content}} {{- else -}} {{ inline .Solution.Content }} {{ end }}",
-		Active: "* {{ if .Code.Content -}} {{`code:` | bold | green}} {{ inline .Code.Content | bold}} {{ else }} " +
-			"{{`solution:` | bold | yellow }} {{ inline .Solution.Content | bold }} {{ end }}",
-		Inactive: "{{ if .Code.Content -}} {{`code:` | green }} {{ inline .Code.Content }} " +
-			"{{- else -}} {{`solution:` | yellow}} {{ inline .Solution.Content }} {{ end }}",
-		Selected: " {{ `✓` | green }} {{if .Code.Content -}} {{ inline .Code.Content | bold }} {{- else -}} {{ inline .Solution.Content | bold }} {{ end }}",
-		Details: "Type: {{- if .Code.Content }} code {{ else }} solution {{- end }}" +
-			"{{ if .Code.Alias }} | Alias: {{ .Code.Alias }} {{- end }}" +
-			"{{ if .Comment }} | Comment: {{ .Comment }} {{- end }}",
-		FuncMap: funcMap,
-	}
-}
-
-func extractScriptStringContent(script store.Script) string {
-	code := strings.Replace(strings.ToLower(script.Code.Content), " ", "", -1)
-	solution := strings.Replace(strings.ToLower(script.Solution.Content), " ", "", -1)
-	comment := strings.Replace(strings.ToLower(script.Comment), " ", "", -1)
-	alias := strings.Replace(strings.ToLower(script.Code.Alias), " ", "", -1)
-	content := fmt.Sprintf("%s %s %s %s", code, solution, comment, alias)
-	return content
-}
-
-func doesScriptContain(script store.Script, input string) bool {
-	input = strings.TrimSpace(strings.ToLower(input))
-
-	//Build pattern
-	pattern := ".*" + strings.Join(strings.Split(input, " "), ".*")
-
-	matched, err := regexp.Match(pattern, []byte(extractScriptStringContent(script)))
-	if err != nil {
-		return false
-	}
-
-	return matched
 }
 
 func findCommand(cmd *cobra.Command, args []string, conf config.Config) error {
@@ -120,21 +63,8 @@ func findCommand(cmd *cobra.Command, args []string, conf config.Config) error {
 	}
 
 	scripts := storeData.Scripts
-	searchScript := func(input string, index int) bool {
-		s := scripts[index]
-		return doesScriptContain(s, input)
-	}
 
-	prompt := promptui.Select{
-		Label:             "Enter your search text",
-		Items:             scripts,
-		Size:              selectItemsSize,
-		StartInSearchMode: true,
-		Searcher:          searchScript,
-		Templates:         getTemplates(),
-	}
-
-	_, result, err := prompt.Run()
+	_, result, err := printers.SelectScriptEntry(scripts)
 	if err != nil {
 		return fmt.Errorf("prompt failed %v", err)
 	}
